+++ modules/hanna-react/CHANGELOG.md
@@ -4,6 +4,7 @@
 
 - ... <!-- Add new lines here. -->
 - feat: Add optional `altText` parameter to `useGetSVGtext()`
+- feat: Export `SSRSupportProps` type
 - fix: `Layout` components set `alt="Reykjavík"` text on their header logo
 - fix: Raise `Multiselect`'s internal `summaryLimit` to `20`
 - fix: List `tslib` as dependency